In 64 bit build size_t is "unsigned long long" thus strings can actually be longer then 2 giga characters. It is not an issue for UI strings thus I've just casted strlen() and alike to "int"

+++ glui_filebrowser.cpp
@@ -116,7 +116,7 @@ void GLUI_FileBrowser::fbreaddir(const GLUI_String &d) {
     list->delete_all();
     if (hFind != INVALID_HANDLE_VALUE) {
       do {
-        int len = strlen(FN.cFileName);
+        int len = (int)strlen(FN.cFileName);
         if (FN.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY) {
           item = '\\';
           item += FN.cFileName;
